+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [3.0.0] - 2019-11-29
 ### Added
 - Support for webp responses under the `ImageFormatter` error handler
+- Additional XML content types [#9]
 
 ### Changed
 - Merged `JpegFormatter`, `GifFormatter` and `PngFormatter` in one `ImageFormatter`.
@@ -102,6 +103,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.1.0 - 2016-10-03
 First version
+
+[#9]: https://github.com/middlewares/error-handler/issues/9
 
 [3.0.0]: https://github.com/middlewares/error-handler/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/middlewares/error-handler/compare/v1.2.0...v2.0.0