+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.1] - 2020-12-03
+### Added
+- Support for PHP 8.0
+
 ## [3.0.0] - 2019-11-29
 ### Added
 - Support for webp responses under the `ImageFormatter` error handler
@@ -106,6 +110,7 @@ First version
 
 [#9]: https://github.com/middlewares/error-handler/issues/9
 
+[3.0.1]: https://github.com/middlewares/error-handler/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/middlewares/error-handler/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/middlewares/error-handler/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/middlewares/error-handler/compare/v1.1.0...v1.2.0